Refactor payment detail information to its own service under common and refactor giving history and recurring giving lists to use this service.
