+++ crossroads.net/app/common/giving/services/index.js
@@ -9,6 +9,7 @@
     .factory('Programs', require('./programs.service'))
     .factory('DonationService', require('./donation.service'))
     .factory('GiveFlow', require('./giveFlow.service'))
+    .factory('PaymentDisplayDetailService', require('./paymentDisplayDetail.service'))
     .constant('CC_BRAND_CODES', require('./ccBrand.constants'))
     ;
 })();
