+++ crossroads.net/app/profile/giving/recurring/recurring_giving_list.directive.js
@@ -3,9 +3,9 @@
 
   module.exports = RecurringGivingList;
 
-  RecurringGivingList.$inject = ['$log'];
+  RecurringGivingList.$inject = ['$log', 'PaymentDisplayDetailService'];
 
-  function RecurringGivingList($log) {
+  function RecurringGivingList($log, PaymentDisplayDetailService) {
     return {
       restrict: 'EA',
       transclude: true,
@@ -18,49 +18,8 @@
 
     function link(scope) {
       scope.$watch('recurringGiftsInput', function(recurringGifts) {
-        scope.recurringGifts = postProcessRecurringGifts(recurringGifts);
+        scope.recurringGifts = PaymentDisplayDetailService.postProcess(recurringGifts);
       });
-    }
-
-    function getCardIcon(brand) {
-      switch (brand) {
-        case 'Visa':
-          return ('cc_visa');
-        case 'MasterCard':
-          return ('cc_mastercard');
-        case 'Discover':
-          return ('cc_discover');
-        case 'AmericanExpress':
-          return ('cc_american_express');
-        default:
-          return ('');
-      }
-    }
-
-    function postProcessRecurringGifts(recurringGiftsInput) {
-      var recurringGifts = _.transform(recurringGiftsInput, function(result, rg) {
-          var recurringGift = _.cloneDeep(rg);
-          setRecurringGiftDisplayDetails(recurringGift.source);
-          result.push(recurringGift);
-        });
-
-      return recurringGifts;
-    }
-
-    function setRecurringGiftDisplayDetails(source) {
-      switch (source.type) {
-        case 'Bank':
-        case 'Check':
-          source.icon = 'library';
-          source.viewBox = '0 0 32 32';
-          source.name = 'ending in ' + source.last4;
-          break;
-        case 'CreditCard':
-          source.icon = getCardIcon(source.brand);
-          source.viewBox = '0 0 160 100';
-          source.name = 'ending in ' + source.last4;
-          break;
-      }
     }
   }
 })();
